Pull host output dir tests out to separate file
