+++ src/NServiceBus.Core.Tests/Logging/DefaultFactoryTests.cs
@@ -18,14 +18,6 @@
             Assert.Throws<ArgumentNullException>(() => defaultFactory.Directory(""));
             Assert.Throws<ArgumentNullException>(() => defaultFactory.Directory(" "));
         }
-
-        [Test]
-        public void When_not_running_ASP_NET_should_choose_BaseDirectory_as_logging_directory()
-        {
-            var directory = Host.GetOutputDirectory();
-
-            Assert.AreEqual(AppDomain.CurrentDomain.BaseDirectory, directory);
-        }
     }
